Remove CHANGELOG.md from WpKit theme
The CHANGELOG.md file was deleted from the WpKit theme to simplify the project's file structure. Having versioning history inside the codebase seemed redundant and a bit confusing as the exact same information is already documented and accessible directly on the project's public GitHub repository. This change should not impact the theme functionality or versioning whatsoever.