Update and refine wp_enqueue_script and wp_enqueue_style calls
The 'wpkit-script' has been renamed to 'wpkit-frontend' and the references in wp_enqueue_script and wp_enqueue_style have been updated. The 'wpkit-style' call to get_stylesheet_uri with the current theme version added as a parameter to ensure the most recent stylesheet is used. Also, the 'enqueue_block_editor_assets' function is updated for better readability.