Update README.md with more detailed theme installation process
This commit updates the installation instructions of the WpKit WordPress theme in the README.md file and provides clearer instructions. It adds is a step-by-step guide on downloading the theme from the GitHub repository, and properly installing it into the WordPress theme directory. This should help new users and less tech-savvy individuals to easily set up the theme.