Add new theme.scss file to WPKit theme
A new scss file named theme.scss has been added to the WPKit theme directory. The file implements the use of a prefix for WPKit and imports critical parts of Bootstrap to enhance the theme styling. The file was added as part of an effort to optimize and streamline the WPKit theme.